requirements.txt
streamlit
PyMuPDF
python-docx
html2text
openai
langchain>=0.1.14
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dotenv
sentence-transformers
transformers
accelerate
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

